+++ cmanpage.c
@@ -70,7 +70,7 @@ main(int argc, char *argv[])
 	/* Read in all of our files now so we can repledge. */
 
 	if (argc > 0 &&
-	    (confs = calloc(argc, sizeof(FILE *))) == NULL)
+	    (confs = calloc((size_t)argc, sizeof(FILE *))) == NULL)
 		err(1, NULL);
 
 	for (i = 0; i < (size_t)argc; i++)
